[*]вынесла реализацаю isInt в cpp файл
