+++ sources/im3_im4/is_int.cpp
@@ -1,0 +1,11 @@
+#include "is_int.h"
+
+bool isInt(const std::string &nameVar)
+{
+    for (auto &letter : nameVar) {
+        if (!isdigit(letter)) {
+            return false;
+        };
+    }
+    return true;
+}
