+++ sources/im3_im4/is_int.h
@@ -1,11 +1,5 @@
+#pragma once
+#include <string>
 #include <cctype>
 
-bool isInt(const std::string &nameVar)
-{
-    for (auto &letter : nameVar) {
-        if (!isdigit(letter)) {
-            return false;
-        };
-    }
-    return true;
-}
+bool isInt(const std::string &nameVar);
